chore: Move validation sets endpoints to the endpoints section of app
